requirements.txt
bcrypt==3.2.0
fastapi==0.75.0
PyJWT==2.3.0
passlib==1.7.4
tortoise-orm==0.19.0
uvicorn==0.17.6